Experimental mixin optimisation
I need to take a lot more time reviewing and simplifying this code
before I'm happy with it, but it does something pretty powerful.  It
makes mixins and mixin blocks be functions that return strings.  It
then hoists all mixins that don't refer to any locals/global variables.
 Finally, for all mixins where there is a 1-1 mapping between mixin
name and code, it updates the reference to directly refer to the
function, rather than going via the jade_mixins object.